Added Draft List
This keeps a record of people's draft edits in case they forget about them.
It adds a 'Drafts' link to the Find menu to show them.
It will show drafts that have been made after installing this feature.
I also added a class to the body to show other apps (specifically WT BEE) that this browser has this extension installed.