Update/Change Image Zoom
Hover over an image. The image zoom feature is activated when you hover over an image. You'll see a "+" sign inside a circle appear on the bottom right corner of the image.

Click the "+". Click the "+" sign to zoom in on the image. The image will expand to a larger size, and the rest of the page will darken to focus on the zoomed image.

Move the image. You can click and drag the zoomed image around the screen to reposition it.

Zoom in and out on the zoomed image. Use your mouse wheel to further zoom in or out on the zoomed image. If you don't have a mouse wheel, click on the zoomed image to zoom in.

Close the zoomed image. You can close the zoomed image in two ways:

Double click on the zoomed image.
Click anywhere on the darkened part of the page.
Remember, the zoomed image will reset back to its original size and position when you move your mouse away from the image and then hover over it again.