Cleanup the CC7 deltas DB
We remove empty deltas and ones older than 45 days (unless they are
the most recent ones for their user). We also show all deltas
(previously ones with >500 additions were ignored), but will
only show the first 500 items in each.

Future enhancement: provide the user with a way to see all of
a large set (perhaps an additional pop-up that only shows it).