Make ID retrieval more reliable
Replace all retrievals of cookies wikitree_wtb_UserName and wikitree_wtb_UserID with calls to 2 common functions that get the 2 ids from menu items.  This should be more reliable in the face of the cookies disappearing at times.

This also removes addLogInLogOutMessage as, hopefully, this should no longer be necessary.